Library

Library

Parameters

Problems

Continuation

Utils

Misc.

Parameters

Problems

HclinicBifurcationKit.HomoclinicHyperbolicProblemPBC — Type

Computation of homoclinic orbit to an hyperbolic saddle based on the projection boundary condition (PBC) method.

bvp::Any: Sructure encoding the boundary value problem. For example, you can pass a PeriodicOrbitTrapProblem, a PeriodicOrbitOCollProblem or an AbstractShootingProblem

lens::Any: Two lenses which are used to define 2 free parameters.

T::Any: Return time T

ϵ0::Any: Precision of how far the section is from the homoclinic point

ϵ1::Any: Precision of how far the section is from the homoclinic point

freelens::Any: Free parameters

Qu0::Any: Orthonormal Projector on the unstable subspace orthogonal

Qs0::Any: Orthonormal Projector on the stable subspace orthogonal

N::Int64: Dimension of phase space

updateEveryStep::Int64: updates the section every update_section_every_step step during continuation

jacobian::Symbol: How the jacobian of the problem is computed

test::Any

testOrbitFlip::Bool

testInclinationFlip::Bool

nUnstable::Int64

nStable::Int64

Continuation

BifurcationKit.continuation — Method

continuation(
    𝐇𝐨𝐦,
    homguess,
    lens,
    alg,
    _contParams;
    plot_solution,
    kwargs...
)

This is the continuation method for computing a homoclinic solution to a hyperbolic saddle. The parameter lens is the one from prob_vf::BifurcationProblem.

Arguments

Similar to continuation except that the problem is a HomoclinicHyperbolicProblemPBC.

BifurcationKit.continuation — Method

continuation(
    prob_vf,
    bt,
    bvp,
    alg,
    _contParams;
    ϵ0,
    amplitude,
    freeparams,
    maxT,
    update_every_step,
    test_orbit_flip,
    test_inclination_flip,
    kwargs...
)

Perform automatic branch switching to homoclinic curve from a Bogdanov-Takens bifurcation point. It uses the homoclinic orbit predictor from the Bogdanov-Takens normal form.

Arguments

prob::BifurcationProblem contains the vector field

bt::BK.BogdanovTakens a Bogdanov-takens point. For example, you can get this from a call to bt = get_normal_form(br, ind_bt)

bvp::BK.AbstractBoundaryValueProblem, for example PeriodicOrbitOCollProblem(50, 4)

alg continuation algorithm

_contParams::ContinuationPar

Optional arguments

ϵ0 = 1e-5 distance of the homolinic orbit from the saddle point

amplitude = 1e-3 amplitude of the homoclinic orbit

maxT = Inf limit on the "period" of the homoclinic cycle.

You can also pass the same arguments to the constructor of ::HomoclinicHyperbolicProblemPBC and those to continuation from BifurcationKit.

kwargs arguments passed to continuation

Utils

HclinicBifurcationKit.generate_hom_problem — Function

generate_hom_problem(
    coll,
    x,
    pars,
    lensHom;
    verbose,
    ϵ0,
    ϵ1,
    t0,
    t1,
    maxT,
    freeparams,
    kw...
)

Generate a homoclinic to hyperbolic saddle problem from a periodic solution obtained with problem pb.

Arguments

coll a PeriodicOrbitOCollProblem which provide basic information, like the number of time slices M

x::AbstractArray initial guess

pars parameters

lensHom::Lens parameter axis for continuation

ϵ0, ϵ1: specify the distance to the saddle point of x₀, x₁

t0, t1: specify the time corresponding to x₀, x₁. Overwrite the part with ϵ0, ϵ1 if set.

Optional arguments

You can pass the same arguments to the constructor of ::HomoclinicHyperbolicProblemPBC.

Output

returns a HomoclinicHyperbolicProblemPBC and an initial guess.

generate_hom_problem(
    sh,
    x,
    pars,
    lensHom;
    verbose,
    time,
    ϵ0,
    ϵ1,
    t0,
    t1,
    maxT,
    freeparams,
    kw...
)

Generate a homoclinic to hyperbolic saddle problem from a periodic solution obtained with problem pb.

Arguments

sh a ShootingProblem which provide basic information, like the number of time slices M

x::AbstractArray initial guess

pars parameters

lensHom::Lens parameter axis for continuation

ϵ0, ϵ1: specify the distance to the saddle point of x₀, x₁

t0, t1: specify the time corresponding to x₀, x₁. Overwrite the part with ϵ0, ϵ1 if set.

Optional arguments

You can pass the same arguments to the constructor of ::HomoclinicHyperbolicProblemPBC.

Output

returns a HomoclinicHyperbolicProblemPBC and an initial guess.

Missing docstring.

Missing docstring for get_homoclinic_orbit. Check Documenter's build log for details.

Misc.